Move some action tests to test_models
